@@ -4,6 +4,11 @@ import './index.css';
 import App from './App';
 
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(
+<App />
+// <App/>
+)
+// ReactDOM.render(<App />, document.getElementById('root'));
 
 
